refactor: folders dynamo db
